Resolve the rest of the TODOs
- Password Reset flow change: password reset now invalidates (delete) existing unused tokens before granting a new one.
- Updated JWT token claim to use i64 (timestamp native) and add validation.
- Ignored rate_limiter todo

+++ src/models/auth.rs
@@ -15,17 +15,20 @@ pub struct MessageResponse<T> {
 // JWT Claims
 // ============================================================================
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Validate)]
 pub struct Claims {
     pub sub: String, // user_id
-    pub exp: usize,  // expiration (Unix timestamp)
-    pub iat: usize,  // issued at (Unix timestamp)
+    #[validate(range(min = 0, message = "Timestamp must be a positive integer"))]
+    pub exp: i64, // expiration (Unix timestamp)
+    #[validate(range(min = 0, message = "Timestamp must be a positive integer"))]
+    pub iat: i64, // issued at (Unix timestamp)
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Validate)]
 pub struct EmailVerifyClaims {
     pub sub: Uuid,
-    pub exp: usize,
+    #[validate(range(min = 0, message = "Timestamp must be a positive integer"))]
+    pub exp: i64,
     pub purpose: String,
 }
